organize shell code into its own file
